Moved methods for reversal from decryption, to the rotation module.
