+++ lib/decryption.rb
@@ -1,9 +1,10 @@
 require 'date'
 require_relative 'generator'
-require_relative 'enigma'
+require_relative 'rotation'
 
-class Decryption 
+class Decryption
   include Generator
+  include Rotation
 
     attr_reader :a_key,
                 :b_key,
@@ -31,19 +32,19 @@ class Decryption
     fourth_reversal.join("").to_s
   end
 
-  def first_reversal
-    generate_reversal(@a_key, 0)
-  end
-
-  def second_reversal
-    generate_reversal(@b_key, 1)
-  end
-
-  def third_reversal
-    generate_reversal(@c_key, 2)
-  end
-
-  def fourth_reversal
-    generate_reversal(@d_key, 3)
-  end
+  # def first_reversal
+  #   generate_reversal(@a_key, 0)
+  # end
+  #
+  # def second_reversal
+  #   generate_reversal(@b_key, 1)
+  # end
+  #
+  # def third_reversal
+  #   generate_reversal(@c_key, 2)
+  # end
+  #
+  # def fourth_reversal
+  #   generate_reversal(@d_key, 3)
+  # end
 end
