+++ lib/encryption.rb
@@ -1,8 +1,7 @@
 require 'date'
 require_relative 'generator'
-require_relative 'enigma'
 
-class Encryption 
+class Encryption
 include Generator
 
   attr_reader :a_key,
